(feature/map): remove unused code
